pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "translategemma-cli"
version = "0.2.0"
description = "Local CLI for TranslateGemma translation (multi-platform, multi-model)"
readme = "README.md"
requires-python = ">=3.11"
license = {text = "MIT"}
authors = [
    {name = "Jacky Chan"}
]
keywords = ["translation", "cantonese", "mlx", "apple-silicon", "gemma", "pytorch", "cuda"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Environment :: Console",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Operating System :: MacOS",
    "Operating System :: POSIX :: Linux",
    "Operating System :: Microsoft :: Windows",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Text Processing :: Linguistic",
]
dependencies = [
    "transformers>=4.42.0,<5.0",
    "huggingface-hub>=0.24.0,<1.0",
    "typer>=0.12.0,<1.0",
    "pyyaml>=6.0,<7.0",
    "rich>=13.0.0,<15.0",
    "prompt-toolkit>=3.0.0,<4.0",
    "regex>=2024.0.0",
    "accelerate>=0.25.0,<1.0",
]

[project.optional-dependencies]
# Apple Silicon (MLX backend)
mlx = [
    "mlx>=0.22.0",
    "mlx-lm>=0.20.0",
]
# NVIDIA GPU (PyTorch + CUDA)
cuda = [
    "torch>=2.0.0,<3.0",
    "bitsandbytes>=0.41.0,<1.0",
]
# CPU-only (PyTorch)
cpu = [
    "torch>=2.0.0,<3.0",
]
# Development dependencies
dev = [
    "pytest>=7.0.0,<9.0",
    "pytest-cov>=4.0.0,<8.0",
    "black>=23.0.0",
    "ruff>=0.1.0",
]

[project.scripts]
translate = "translategemma_cli.cli:app"

[project.urls]
"Homepage" = "https://github.com/jhkchan/translategemma-cli"

[tool.setuptools.packages.find]
where = ["."]
include = ["translategemma_cli*"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = "-v --tb=short"

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.black]
line-length = 100
target-version = ["py311"]
